Relocated source files to a package folder for release
